pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kinetic-ledger"
version = "0.1.0"
description = "Trust-minimized motion commerce stack (Phase-2 demo)"
authors = [{ name = "Kinetic Ledger Team" }]
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.115.0",
    "pydantic>=2.7.0",
    "uvicorn>=0.30.0",
    "eth-utils>=2.3.0",
    "numpy>=1.24.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
    "google-generativeai>=0.3.0",  # Gemini API for analysis + embeddings + File Search
    "eth-account>=0.10.0",
    "qdrant-client>=1.7.0",
    "web3>=6.0.0",
    "matplotlib>=3.7.0",  # Frame rendering for Gemini analysis
    "Pillow>=10.0.0",  # Image processing for frame rendering
]

[project.optional-dependencies]
warehouse = ["google-cloud-storage>=2.14.0"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21.0", "black>=23.0", "ruff>=0.1.0", "mypy>=1.0"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
strict = true
